Move coffee list related files to the pages folder
