+++ src/App.js
@@ -3,7 +3,7 @@ import { Redirect, Route, Switch } from 'react-router-dom'
 import { Container } from 'semantic-ui-react'
 import { graphql } from 'react-apollo'
 import LeanCoffeeDetails from './components/coffees/Details/Details'
-import ListLeanCoffees from './components/coffees/List'
+import { ListLeanCoffees } from './pages'
 import ListParticipants from './components/participants/List'
 import ListTopics from './components/topics/List'
 import Home from './components/Home'
