+++ src/pages/index.js
@@ -1,0 +1,1 @@
+export { default as ListLeanCoffees } from './coffee/list/List'
